split namable decorator into separate files
